terminando função de cadastro de pedidos e de lanches

+++ src/Lanche.java
@@ -1,13 +1,13 @@
 public class Lanche {
   private int codigo;
   private String nome;
-  private float valor;
-  
+  private double valor;
+
   public Lanche() {
     
   }
   
-  public Lanche(int codigo, String nome, float valor) {
+  public Lanche(int codigo, String nome, double valor) {
     this.codigo = codigo;
     this.nome = nome;
     this.valor = valor;
@@ -21,7 +21,7 @@ public class Lanche {
     return this.nome;
   }
   
-  public float getValor() {
+  public double getValor() {
     return this.valor;
   }
 }
